Refactor : 스크린 별 패키지 분리
1. 스크린 별 탑 바 구현
2. 패키지 명 수정(첫 대문자 -> 첫 소문자)
3. AppBar.kt 삭제
4. 오토 커벤션
